added runscript API call cleaned rts2script library
git-svn-id: http://svn.code.sf.net/p/rts-2/code/trunk/rts-2@9673 65ab8f4e-f147-0410-b3a9-f14133ecfe55

+++ lib/xmlrpc++/XmlRpcClient.h
@@ -130,7 +130,7 @@ namespace XmlRpc
 			int _sendAttempts;
 
 			// Number of bytes of the request that have been written to the socket so far
-			int _bytesWritten;
+			size_t _bytesWritten;
 
 			// Non-zero (no IDLE) if some request (either XML-RPC or HTTP GET) is executed. If you want to multithread,
 			// each thread should have its own client.
